bug #278: enhance tree management component
- Update trees component with improved data handling
- Add tab-based interface for better organization
- Fix styling issues in tree form component

+++ src/app/components/trees/trees.component.ts
@@ -2,8 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {TreeService, Tree} from '../../frontend-core/src/lib/public_api';
 import {UtilsService} from '../../services/utils.service';
 import {Router} from '@angular/router';
-import {config} from 'src/config';
-import {Observable, Subject} from 'rxjs';
+import {Observable, Subject, firstValueFrom} from 'rxjs';
 import {MatDialog} from '@angular/material/dialog';
 import {DialogMessageComponent} from '../../frontend-gui/src/lib/public_api';
 
@@ -15,25 +14,21 @@ import {DialogMessageComponent} from '../../frontend-gui/src/lib/public_api';
 export class TreesComponent implements OnInit {
   saveAgGridStateEvent: Subject<boolean> = new Subject<boolean>();
   dataUpdatedEvent: Subject<boolean> = new Subject<boolean>();
-  themeGrid: any = config.agGridTheme;
   columnDefs: any[];
   gridModified = false;
 
-  constructor(public dialog: MatDialog,
-              public treeService: TreeService,
-              private utils: UtilsService,
-              private router: Router,
-  ) {
-
-  }
+  constructor(
+    public dialog: MatDialog,
+    public treeService: TreeService,
+    private utils: UtilsService,
+    private router: Router,
+  ) {}
 
   ngOnInit() {
-
-    var columnEditBtn = this.utils.getEditBtnColumnDef();
+    const columnEditBtn = this.utils.getEditBtnColumnDef();
     columnEditBtn['cellRendererParams'] = {
       clicked: this.newData.bind(this)
     };
-
 
     this.columnDefs = [
       columnEditBtn,
@@ -41,79 +36,68 @@ export class TreesComponent implements OnInit {
       this.utils.getIdColumnDef(),
       this.utils.getEditableColumnDef('treesEntity.name', 'name')
     ];
-
   }
 
-  async canDeactivate(): Promise<boolean | Observable<boolean>> {
-
-    if (this.gridModified) {
-
-
-      const result = await this.utils.showNavigationOutDialog().toPromise();
-      if (!result || result.event !== 'Accept') {
-        return false;
-      } else if (result.event === 'Accept') {
-        return true;
-      } else {
-        return true;
-      }
-    } else {
+  async canDeactivate(): Promise<boolean> {
+    if (!this.gridModified) {
       return true;
     }
+
+    const result = await firstValueFrom(this.utils.showNavigationOutDialog());
+    return !result || result.event === 'Accept';
   }
 
-  setGridModifiedValue(value) {
+  setGridModifiedValue(value: boolean): void {
     this.gridModified = value;
   }
 
-  getAllTrees = () => {
-
+  getAllTrees = (): Observable<Tree[]> => {
     return this.treeService.getAll();
   };
 
-  newData(id: any) {
+  async newData(id: any): Promise<void> {
     this.saveAgGridStateEvent.next(true);
-    this.router.navigate(['trees', id, 'treesForm']);
+    await this.router.navigate(['trees', id, 'treesForm']);
   }
 
-  applyChanges(data: Tree[]) {
+  async applyChanges(data: Tree[]): Promise<void> {
     const promises: Promise<any>[] = [];
+
     data.forEach(tree => {
       promises.push(new Promise((resolve) => {
         this.treeService.update(tree).subscribe(() => {
           resolve(true);
         });
       }));
-      Promise.all(promises).then(() => {
-        this.dataUpdatedEvent.next(true);
-      });
     });
+
+    await Promise.all(promises);
+    this.dataUpdatedEvent.next(true);
   }
 
-  add(data: Tree[]) {
-    this.router.navigate(['trees', -1, 'treesForm', data[0].id]);
+  async add(data: Tree[]): Promise<void> {
+    await this.router.navigate(['trees', -1, 'treesForm', data[0].id]);
   }
 
-  removeData(data: Tree[]) {
-
+  removeData(data: Tree[]): void {
     const dialogRef = this.dialog.open(DialogMessageComponent);
     dialogRef.componentInstance.title = this.utils.getTranslate('caution');
     dialogRef.componentInstance.message = this.utils.getTranslate('removeMessage');
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        if (result.event === 'Accept') {
-          const promises: Promise<any>[] = [];
-          data.forEach(tree => {
-            promises.push(new Promise((resolve,) => {
-              this.treeService.delete(tree).subscribe(() => {
-                resolve(true);
-              });
-            }));
-            Promise.all(promises).then(() => {
-              this.dataUpdatedEvent.next(true);
+
+    dialogRef.afterClosed().subscribe(async result => {
+      if (result && result.event === 'Accept') {
+        const promises: Promise<any>[] = [];
+
+        data.forEach(tree => {
+          promises.push(new Promise((resolve) => {
+            this.treeService.delete(tree).subscribe(() => {
+              resolve(true);
             });
-          });
-        }
+          }));
+        });
+
+        await Promise.all(promises);
+        this.dataUpdatedEvent.next(true);
       }
     });
   }
